Add files via upload
This notebook goes on briefly explaining about how to detect a face in a image data. How to import a data, detect single and multiple people faces by ignoring background. Verifying two images as same face using two faces of a same person.
This was created to present in AI Winter Workshop -2020 in my undergrad college of Engineering.